fixed side x/y offsets to be 16 bit instead of 8 bit (data didnt fit in 8 bit) and moved some memory regions around to make things fit. in order to make some sprite size/offset fields support page out, player sprites vissprites are now calculated immediately after bsp code and before plane code

+++ doomtype.h
@@ -38,7 +38,7 @@ typedef unsigned long long		uint64_t;
 typedef int16_t filehandle_t;
 typedef uint16_t filelength_t;
 typedef int32_t ticcount_t;
-typedef uint8_t texsize_t;
+typedef uint16_t texsize_t;
 
 
 #ifndef __BYTEBOOL__
